llama3_cypher_expert/requirements_api.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
transformers>=4.40.0
torch>=2.0.0
peft>=0.10.0
accelerate>=0.27.0
google-cloud-storage>=2.10.0
pydantic>=2.0.0
